Add plotting of origin uncertainties in web gis
This consists of changes in multiple places:
 - during QuakeML document indexing
    - horizontal uncertainties are parsed and added to the indexes
    - django geometry field now is a list of a Point and a
      MultiLineString (which is empty in case of no uncertainties given)
    - to calculate the end points of the error bar lines, package geopy
      is added as a new dependency (it's available on current Debian
      stable through packaging and also in conda, so no big deal) to do
      the Vincenty calculations
 - in the WebGIS adds a switch to turn on/off uncertainty plotting (as
   it can be quite a haystack for big event clusters)

These changes can be added to existing live Jane instances without
problems, I think. Only a reindexing of all events will be necessary.

I confirmed that it (still) works for Events that don't have an Origin,
or have an Origin without uncertainty information